+++ docs/changes.rst
@@ -22,6 +22,7 @@ Weblate 5.14
 * Improved activity logging for :ref:`addon-weblate.consistency.languages` add-on.
 * :ref:`check-rst-references` and :ref:`check-rst-syntax` now support substitutions.
 * :ref:`component-hide_glossary_matches` component setting.
+* :ref:`file_format_params` can now be configured via the :ref:`api`.
 
 .. rubric:: Bug fixes